chore(team): split step progress paths
